requirements.txt
pip==8.1.2
Flask==0.10.1
PyMySQL3
configparser